# docs/openapi.yaml
openapi: 3.0.3
info:
  title: KORION Chong API
  version: 0.1.0
paths:
...
  /api/partner/dashboard:
    get:
      summary: Sales partner dashboard summary page data
      parameters:
        - $ref: '#/components/parameters/PartnerIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
        - $ref: '#/components/parameters/PeriodQuery'
      responses:
        '200':
          $ref: '#/components/responses/PartnerPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/partner/merchant-applications:
    get:
      summary: Merchant application summary for a sales partner
      description: Returns only partner-owned merchant signup application rows using distributor_signup_applications.owner_partner_id as the server-side scope boundary.
      parameters:
        - $ref: '#/components/parameters/PartnerIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
      responses:
        '200':
          $ref: '#/components/responses/PartnerPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/partner/merchant-applications/detail:
    get:
      summary: Merchant application detail page shell for a sales partner
      parameters:
        - $ref: '#/components/parameters/PartnerIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
      responses:
        '200':
          $ref: '#/components/responses/PartnerPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/partner/merchants:
    get:
      summary: Merchants directly owned by a sales partner
      parameters:
        - $ref: '#/components/parameters/PartnerIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
      responses:
        '200':
          $ref: '#/components/responses/PartnerPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/partner/merchant-sales:
    get:
      summary: Sales summary for merchants directly owned by a sales partner
      parameters:
        - $ref: '#/components/parameters/PartnerIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
        - $ref: '#/components/parameters/PeriodQuery'
      responses:
        '200':
          $ref: '#/components/responses/PartnerPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/partner/settlements/request-summary:
    get:
      summary: Sales partner settlement request summary page data
      parameters:
        - $ref: '#/components/parameters/PartnerIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
        - $ref: '#/components/parameters/PeriodQuery'
      responses:
        '200':
          $ref: '#/components/responses/PartnerPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/partner/settlements:
    get:
      summary: Sales partner settlement history page data
      parameters:
        - $ref: '#/components/parameters/PartnerIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
      responses:
        '200':
          $ref: '#/components/responses/PartnerPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/partner/settlements/detail:
    get:
      summary: Sales partner settlement detail page data
      parameters:
        - $ref: '#/components/parameters/PartnerIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
      responses:
        '200':
          $ref: '#/components/responses/PartnerPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/partner/settlements/requests:
    post:
      summary: Create a sales partner settlement request
      description: Recalculates available commission entries server-side and creates a REQUESTED distributor_settlement_requests row for the authenticated sales partner.
      parameters:
        - $ref: '#/components/parameters/PartnerIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettlementCreateRequest'
      responses:
        '200':
          description: Settlement request created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettlementActionResponse'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/partner/hq-notices:
    get:
      summary: HQ notices visible to the sales partner
      parameters:
        - $ref: '#/components/parameters/PartnerIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
      responses:
        '200':
          $ref: '#/components/responses/PartnerPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/partner/notices/send-summary:
    get:
      summary: Sales partner notice send recipient summary
      parameters:
        - $ref: '#/components/parameters/PartnerIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
      responses:
        '200':
          $ref: '#/components/responses/PartnerPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/partner/notices:
    get:
      summary: Sales partner notice send history page data
      parameters:
        - $ref: '#/components/parameters/PartnerIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
      responses:
        '200':
          $ref: '#/components/responses/PartnerPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/partner/notices/history:
    get:
      summary: Alias for sales partner notice send history page data
      parameters:
        - $ref: '#/components/parameters/PartnerIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
      responses:
        '200':
          $ref: '#/components/responses/PartnerPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/partner/notices/outbox:
    post:
      summary: Send or schedule a sales partner notice to owned merchants
      description: Stores the notice body, merchant recipient delivery rows, and activity log in one transaction. scheduledAt creates SCHEDULED/PENDING history; otherwise the notice is stored as SENT/SENT.
      parameters:
        - $ref: '#/components/parameters/PartnerIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoticeSendRequest'
      responses:
        '200':
          description: Notice send request was stored.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoticeSendResponse'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/partner/profile:
    get:
      summary: Sales partner profile page data
      parameters:
        - $ref: '#/components/parameters/PartnerIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
      responses:
        '200':
          $ref: '#/components/responses/PartnerPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/partner/activity-logs:
    get:
      summary: Sales partner activity log page data
      parameters:
        - $ref: '#/components/parameters/PartnerIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
      responses:
        '200':
          $ref: '#/components/responses/PartnerPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
...
components:
  parameters:
...
    PartnerIdHeader:
      name: X-Partner-Id
      in: header
      required: true
      schema:
        type: integer
        format: int64
      description: Temporary sales partner id until JWT auth is wired.
...
    CountryScopesHeader:
      name: X-Country-Scopes
      in: header
      required: true
      schema:
        type: string
      description: Temporary comma-separated country scopes until JWT auth is wired.
    CountryScopeQuery:
      name: countryScope
      in: query
      required: true
      schema:
        type: string
        minLength: 2
        maxLength: 2
      example: KR
    PeriodQuery:
      name: period
      in: query
      required: false
      schema:
        type: string
        pattern: '^\d{4}-\d{2}$'
      example: '2026-06'
...
  responses:
    ForbiddenCountryScope:
      description: Requested countryScope is outside the authenticated role's server-side country scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
...
    PartnerPageData:
      description: Page data shaped for the current KORION PAY partner admin frontend.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PartnerPageDataResponse'
...
  schemas:
...
    PartnerPageDataResponse:
      type: object
      additionalProperties: true
      description: >
        Transitional frontend page payload. Each partner page returns the same top-level
        keys used by its existing JSON fallback, scoped by X-Partner-Id and countryScope.
...
    SettlementCreateRequest:
      type: object
      required:
        - periodStart
        - periodEnd
        - requestedAmount
      properties:
        periodStart:
          type: string
          format: date
        periodEnd:
          type: string
          format: date
        requestedAmount:
          type: number
          minimum: 0.000000000000000001
        walletAddressId:
          type: integer
          format: int64
          nullable: true
        memo:
          type: string
          maxLength: 2000
        requestId:
          type: string
          maxLength: 128
...
    SettlementActionResponse:
      type: object
      properties:
        settlementRequestId:
          type: integer
          format: int64
        requestNo:
          type: string
        recipientType:
          type: string
          enum:
            - LEADER
            - PARTNER
            - MERCHANT
        recipientPartnerId:
          type: integer
          format: int64
          nullable: true
        recipientMerchantId:
          type: integer
          format: int64
          nullable: true
        requestedAmount:
          type: number
        approvedAmount:
          type: number
        heldAmount:
          type: number
        status:
          type: string
          enum:
            - REQUESTED
            - APPROVED
            - REJECTED
            - PAID
        resultCode:
          type: string
          enum:
            - SETTLEMENT_REQUEST_CREATED
            - SETTLEMENT_REQUEST_APPROVED
            - SETTLEMENT_REQUEST_REJECTED
            - SETTLEMENT_REQUEST_PAID
        messageKey:
          type: string
    NoticeRecipientRequest:
      type: object
      required:
        - recipientType
      properties:
        recipientType:
          type: string
          enum:
            - PARTNER
            - MERCHANT
        recipientPartnerId:
          type: integer
          format: int64
          nullable: true
        recipientMerchantId:
          type: integer
          format: int64
          nullable: true
    NoticeSendRequest:
      type: object
      required:
        - title
        - noticeType
        - body
        - channel
        - recipients
      properties:
        title:
          type: string
          maxLength: 255
        noticeType:
          type: string
          maxLength: 64
        body:
          type: string
        channel:
          type: string
          enum:
            - IN_APP
            - EMAIL
            - SMS
            - PUSH
        scheduledAt:
          type: string
          format: date-time
          nullable: true
          description: Must be later than the server clock when provided.
        recipients:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/NoticeRecipientRequest'
        requestId:
          type: string
          maxLength: 128
    NoticeSendResponse:
      type: object
      properties:
        noticeId:
          type: integer
          format: int64
        status:
          type: string
          enum:
            - SENT
            - SCHEDULED
        scheduledAt:
          type: string
          format: date-time
          nullable: true
        sentAt:
          type: string
          format: date-time
          nullable: true
        recipientCount:
          type: integer
        resultCode:
          type: string
          enum:
            - NOTICE_SENT
            - NOTICE_SEND_SCHEDULED
        messageKey:
          type: string
...
    ApiErrorResponse:
      type: object
      required:
        - timestamp
        - code
        - message
      properties:
        timestamp:
          type: string
          format: date-time
        code:
          type: string
        message:
          type: string
...